pyproject.toml
[project]
name = "ai-podcast"
version = "0.1.0"
description = ""
authors = [
    {name = "Sairam M",email = "sairam.m.2005.65@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "google-generativeai (>=0.8.4,<0.9.0)",
    "gtts (>=2.5.4,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pydub (>=0.25.1,<0.26.0)",
    "fastapi[all] (>=0.115.12,<0.116.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
